requirements.txt
matplotlib==3.7.1
pandas==2.0.0
pandas==2.0.2
pythonping==1.1.4
speedtest_cli==2.1.3
